Sarthak | Adds CXX_FLAGS and corrects the issues reported by those flags

+++ src/engine/storage/SkipListInternalNode.cpp
@@ -14,7 +14,7 @@ namespace pmem {
             }
 
             SkipListInternalNode::~SkipListInternalNode() {
-                for (int index = 0; index < this->forwards.size(); index++) {
+                for (size_t index = 0; index < this->forwards.size(); index++) {
                     this->forwards[index] = nullptr;
                 }
                 this->forwards.clear();
